linux-user: Fix signal before blocking system calls race and SA_RESTART
If a signal is delivered immediately before a blocking system calls the
handler will only be called after the system call returns, which may be a
long time later or never.

This is fixed by using a function (safe_syscall_base) that checks  if a guest
signal is pending prior to making a system call, and if so does not call the
system call and returns -TARGET_ERESTARTSYS. If a signal is received between
the check and the system call host_signal_handler() rewinds execution to
before the check. If the system call returns an error a guest error code
is returned.

safe_syscall_base is implemented in assembly language with initially
only a x86-64 version, a fall back is provided keep the old behaviour.

Also setting the SA_RESTART flag would result in host system calls being
restarted and the guest signal handler only being called when they have
completed. This is also fixed.

This commit contains general infrastructure and safe_syscall_base for x86-64.

Signed-off-by: Timothy Edward Baldwin <T.E.Baldwin99@members.leeds.ac.uk>