implement qemu_execve() function
Introduce a new option, -execve, which will intercept all calls to
execve() and inject the intepreter to the call before passing it to the
host kernel. This has the effect of being able to emulate forking
processes without needing help from the host system using binfmt_misc or
some other method.

Currently the path to the emulator is hard-coded, it should probably be
specified in the parameter.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>